review: move assistants list query schema out of route.ts
Next.js route handlers only recognize HTTP-method and route-segment-
config exports; the extra `listQuerySchema` export (added for unit
testing REV-SEC-168) would make `next build` reject the route. Moved
the schema to a sibling query-schema.ts module and updated the test
import.

Refs PR #1135 review comment from chatgpt-codex-connector (P1).
